% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{ColMaker}
\alias{ColMaker}
\title{Makes unique colours for all but the controls.}
\usage{
ColMaker(genos, SetControl = c(`Mb 311` = "#000000"))
}
\arguments{
\item{genos}{a list of genotypes requiring colours.}

\item{SetControl}{Any predefined colours to use this can be '#FF0000' or 'red'.}
}
\value{
A named list of colours with the geno as the name.
}
\description{
The plans need each plot coloured by genotype
}
\examples{
ColMaker(LETTERS[1:12])


}
