% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zigzag-make.R
\name{MakeMainPlan}
\alias{MakeMainPlan}
\title{Makes a set of plots from a list of genotypes.}
\usage{
MakeMainPlan(xopplist, rowse, colse, nr = 1, nc = 1, zigzag = "R",
  origin = "TL", zz = "R", ori = "TL", FlipR = F)
}
\arguments{
\item{xopplist}{is the name of the genotype clone etc.}

\item{rowse}{the columns in the field n:n eg 1 to 7 (1:7) or 1 to 5 (1:5). This should normaly start at 1.}

\item{colse}{the rows in the field n:n eg 1 to 5 (1:5) or 1 to 6 (1:6).  This should normaly start at 1.}

\item{nr}{the number of rows in the sub plots.}

\item{nc}{the number of columns in the sub plots.}

\item{zigzag}{the zig-zag. This can be rows ('R'), columns ('C'), no zig-zag columns ('NC'), or no zig-zag rows ('NR').}

\item{origin}{the origin of the plot so zig-zags are correct. This can be top-left ('TL'), bottom-left ('BL'), top-right ('TR'), or bottom-right ('BR'). Not all are available for all \code{zz} options.}

\item{zz}{the sup plot zig-zag. This can be rows ('R'), columns ('C'), no zig-zag columns ('NC'), or no zig-zag rows ('NR').}

\item{ori}{the sup plot origin of the plot so zig-zags are correct. This can be top-left ('TL'), bottom-left ('BL'), top-right ('TR'), or bottom-right ('BR'). Not all are available for all \code{zz} options.}

\item{FlipR}{if to flip all the row numbers, use if 1,1 is at the bottom right.}
}
\value{
A list of the posishons of the plants \code{xopp} in the \code{colse} and \code{rowse} starting at \code{ori} and zig-zaging according to \code{zz}.
}
\description{
These can include subplots, in doing so this will call \code{\link{MakeSubPlan}} to make the subplots. Any sub plots can have difering zig-zags and origins dfinde in \code{zz} is called by \code{ori}.
}
\examples{
MakeMainPlan(LETTERS[1:12],1:2,1:6,1,1)
}
