% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{MakeSpace}
\alias{MakeSpace}
\title{Makes spaces for the plants.}
\usage{
MakeSpace(pos, BlockFreq, BlockSapcing, PlantSpacing)
}
\arguments{
\item{pos}{Is the dater frame output from \code{\link{MakeMainPlan}} or something with the same hasders.}

\item{BlockFreq}{this is how frequent paths are in the trial.}

\item{BlockSapcing}{This can be a number or a list of numbers the length of the number of paths.}

\item{PlantSpacing}{This can be a number or a list of numbers the length of the rows or columns.}
}
\value{
The spacing of the plant requested
}
\description{
This is only in one dimension.
}
\examples{
MakeSpace(3,2,1,0.66)

}
