% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zigzag-make.R
\name{MakeMainPlan}
\alias{MakeMainPlan}
\title{Makes a set of plots from a list of genotypes.}
\usage{
MakeMainPlan(
  xopplist,
  rowse,
  colse,
  nr = 1,
  nc = 1,
  zigzag = "R",
  origin = "TL",
  zz = "R",
  ori = "TL",
  FlipR = F,
  UID = 1
)
}
\arguments{
\item{xopplist}{is the name of the genotype clone etc.}

\item{rowse}{the columns in the field n:n e.g. 1 to 7 (1:7) or 1 to 5 (1:5). This should typically start at 1.}

\item{colse}{The rows in the field n:n eg 1 to 5 (1:5) or 1 to 6 (1:6).  This should typically start at 1.}

\item{nr}{The number of rows in the sub plots.}

\item{nc}{The number of columns in the sub plots.}

\item{zigzag}{The zig-zag. This can be rows ('R'), columns ('C'), no zig-zag columns ('NC'), or no zig-zag rows ('NR').}

\item{origin}{The origin of the plot so zig-zags are correct. This can be top-left ('TL'), bottom-left ('BL'), top-right ('TR'), or bottom-right ('BR'). Not all are available for all \code{zz} options.}

\item{zz}{The sup plot zig-zag. This can be rows ('R'), columns ('C'), no zig-zag columns ('NC'), or no zig-zag rows ('NR').}

\item{ori}{The sup plot origin of the plot so zig-zags are correct. This can be top-left ('TL'), bottom-left ('BL'), top-right ('TR'), or bottom-right ('BR'). Not all are available for all \code{zz} options.}

\item{FlipR}{If to flip all the row numbers, use if 1,1 is at the bottom right.}

\item{UID}{The UID for the first plant.}
}
\value{
A list of the posishons of the plants \code{xopp} in the \code{colse} and \code{rowse} starting at \code{ori} and zig-zagging according to \code{zz}.
}
\description{
These can include subplots, in doing so this will call \code{\link{MakeSubPlan}} to make the subplots. Any sub plots can have differing zig-zags and origins defined in \code{zz} is called by \code{ori}.
}
\examples{
MakeMainPlan(LETTERS[1:12],1:2,1:6,1,1)

}
