% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zigzag-plot.R
\name{PlotFieldPlan}
\alias{PlotFieldPlan}
\title{Makes a basic diagram from the output of \code{\link{MakeMainPlan}}.}
\usage{
PlotFieldPlan(List, DefaultCols = c(`Mb 311` = "#000000"), SZ = 1.5,
  TD = 0, Label = "geno", FlipR = FALSE)
}
\arguments{
\item{List}{Is the dater frame output from \code{\link{MakeMainPlan}} or something with the same hasders.}

\item{DefaultCols}{Are a list of key value pair sets for any pre-defigned colurs to use.}

\item{SZ}{Is the text sise in the graph this defalts to 1.5 but can be adjusted for large and small field plans.}

\item{TD}{This alows the text in the field plan to be rotated by n degrees (default = 0).}

\item{Label}{The type of labeling to do on the plan. This can be nane/genotype ('geno'), plot number ('number'), location eg. 2-3 ('loc'),  just a dot ('plant'), or the UID/number of the plant ('UID').}

\item{FlipR}{This flips the rows (default = FALSE).}
}
\value{
A \code{ggplot2} heatmap style graph that can be saved with  \code{ggave}.
}
\description{
It is often needed to make minor changes (for dead plants etc.) to the output of \code{\link{MakeMainPlan}} using \code{PlotFieldPlan} is a good way to do that.
}
\examples{
\dontrun{PlotFieldPlan(MakeMainPlan(LETTERS[1:12], 1:3, 1:4), SZ = 5)}
\dontrun{PlotFieldPlan(MakeMainPlan(LETTERS[1:12], 1:3, 1:4,2,3), SZ = 3)}
\dontrun{PlotFieldPlan(MakeMainPlan(LETTERS[1:12], 1:3, 1:4,2,3), SZ = 3, TD = 90, Label ='number')}
\dontrun{PlotFieldPlan(MakeMainPlan(LETTERS[1:12], 1:3, 1:4,2,3), DefaultCols = c("A" = "red", "B" = "green", "C" = "blue"), SZ = 3)}
\dontrun{PlotFieldPlan(MakeMainPlan(LETTERS[1:14], 1:3, 1:5,2,3), DefaultCols = c("A" = "red", "B" = "green", "C" = "blue"), SZ = 3, Label ='UID')}

}
